Fix various issues

+++ src/Util/Logger.php
@@ -44,7 +44,7 @@ class Logger
             return '';
         }
 
-        $time = ((microtime(true) - self::$startTime) * 1000);
+        $time = (int)((microtime(true) - self::$startTime) * 1000);
 
         return sprintf('%02d:%05.02f', floor($time / 60000), ($time % 60000) / 1000);
     }
